@@ -5,6 +5,8 @@ import com.htec.shelffunction.annotation.AuthenticationUser;
 import com.htec.shelffunction.dto.FunctionDTO;
 import com.htec.shelffunction.model.request.CustomFunctionRequestModel;
 import com.htec.shelffunction.model.request.PredefinedFunctionRequestModel;
+import com.htec.shelffunction.model.request.RenameFunctionRequestModel;
+import com.htec.shelffunction.model.request.UpdateCodeFunctionRequestModel;
 import com.htec.shelffunction.model.response.FunctionResponseModel;
 import com.htec.shelffunction.model.response.TextResponseMessage;
 import com.htec.shelffunction.service.FunctionService;
@@ -12,7 +14,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.io.IOException;
 import java.util.List;
 
 @RestController
@@ -21,6 +22,8 @@ public class FunctionController {
 
     public final String FUNCTION_CREATED = "Function created";
     public final String FUNCTION_DELETED = "Function deleted";
+    public final String FUNCTION_NAME_UPDATED = "Function name updated";
+    public final String FUNCTION_CODE_UPDATED = "Function code updated";
 
     private final FunctionService functionService;
 
@@ -34,7 +37,7 @@ public class FunctionController {
 
         functionService.createPredefinedFunction(functionRequestModel, authUser.getId());
 
-        return ResponseEntity.status(HttpStatus.OK).body(new TextResponseMessage(FUNCTION_CREATED, HttpStatus.OK.value()));
+        return ResponseEntity.status(HttpStatus.CREATED).body(new TextResponseMessage(FUNCTION_CREATED, HttpStatus.CREATED.value()));
     }
 
     @PostMapping("/custom")
@@ -43,13 +46,19 @@ public class FunctionController {
 
         functionService.createCustomFunction(customFunctionRequestModel, authUser.getId());
 
-        return ResponseEntity.status(HttpStatus.OK).body(new TextResponseMessage(FUNCTION_CREATED, HttpStatus.OK.value()));
+        return ResponseEntity.status(HttpStatus.CREATED).body(new TextResponseMessage(FUNCTION_CREATED, HttpStatus.CREATED.value()));
     }
 
     @GetMapping
     public ResponseEntity<List<FunctionDTO>> getAllFunctions(@AuthenticationUser AuthUser authUser) {
 
         return ResponseEntity.ok(functionService.getAllFunctionsByUserId());
+    }
+
+    @GetMapping("/{shelfId}/{eventId}")
+    public ResponseEntity<List<Long>> getAllFunctionIdsByShelfIdAndEventId(@PathVariable Long shelfId, @PathVariable Long eventId) {
+
+        return ResponseEntity.ok(functionService.getAllFunctionIdsByShelfIdAndEventId(shelfId, eventId));
     }
 
     @DeleteMapping("/{functionId}")
@@ -64,5 +73,23 @@ public class FunctionController {
     public ResponseEntity<FunctionResponseModel> getFunction(@AuthenticationUser AuthUser authUser, @PathVariable Long functionId) {
 
         return ResponseEntity.ok(functionService.getFunction(functionId, authUser.getId()));
+    }
+
+    @PutMapping("/rename")
+    public ResponseEntity<TextResponseMessage> renameFunction(@AuthenticationUser AuthUser authUser,
+                                                              @RequestBody RenameFunctionRequestModel renameFunctionRequestModel) {
+
+        functionService.updateFunctionName(renameFunctionRequestModel);
+
+        return ResponseEntity.status(HttpStatus.OK).body(new TextResponseMessage(FUNCTION_NAME_UPDATED, HttpStatus.OK.value()));
+    }
+
+    @PutMapping("/code")
+    public ResponseEntity<TextResponseMessage> updateCode(@AuthenticationUser AuthUser user,
+                                                          @RequestBody UpdateCodeFunctionRequestModel updateCodeFunctionRequestModel) {
+
+        functionService.updateFunctionCode(updateCodeFunctionRequestModel, user.getId());
+
+        return ResponseEntity.status(HttpStatus.OK).body(new TextResponseMessage(FUNCTION_CODE_UPDATED, HttpStatus.OK.value()));
     }
 }
