@@ -5,14 +5,14 @@ import { ProgressBarProps } from './progressBar.interfaces';
 export const ProgressBarWrapper = styled.div`
   width: 100%;
   margin-top: ${theme.space.xs};
-  background: ${theme.colors.white};
-  border: 1px solid ${theme.colors.primary};
+  background: ${theme.colors.primary};
+  border: 1px solid ${theme.colors.white};
 `;
 
 export const StyledProgress = styled.div<ProgressBarProps>`
   align-self: flex-start;
-  background: ${theme.colors.primary};
-  color: ${theme.colors.white};
+  background: ${theme.colors.white};
+  color: ${theme.colors.primary};
   font-size: ${theme.fontSizes.xs};
   text-align: center;
   width: ${({ progress }) => `${progress}%`};
